requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.3
boto3==1.28.25
bcrypt==4.0.1
python-dotenv==1.0.0
moto==4.1.0
jinja2==3.1.2
blinker==1.6.2
twilio==8.2.2
